Make functions more usable
