+++ lib/src/presentation/views/home_screen.dart
@@ -1,9 +1,12 @@
 import 'dart:io';
+
 import 'package:flutter/material.dart';
+
 import 'package:loomi_ui_flutter/widgets/custom_button.dart';
+
 import 'package:flutter_leap/main.dart';
-import 'package:flutter_leap/src/presentation/widgets/custom_video_picker.dart';
 import 'package:flutter_leap/src/utils/custom_colors.dart';
+import 'package:flutter_leap/src/utils/helpers/select_videos_sheet_helper.dart';
 import 'package:flutter_leap/src/utils/localization/app_localizations.dart';
 
 import '../widgets/bottom_navigation_bar.dart';
@@ -48,9 +51,43 @@ class _HomeScreenState extends State<HomeScreen> {
                     child: Column(
                       mainAxisAlignment: MainAxisAlignment.center,
                       children: [
-                        CustomVideoPicker(
-                          onAdd: onAdd,
-                          onRemove: onRemove,
+                        GestureDetector(
+                          onTap: () => openSelectVideoSheet(context),
+                          child: Row(
+                            mainAxisSize: MainAxisSize.min,
+                            children: [
+                              Container(
+                                alignment: Alignment.centerLeft,
+                                decoration: BoxDecoration(
+                                  border: Border.all(
+                                    color: Theme.of(context).primaryColor,
+                                  ),
+                                  borderRadius: const BorderRadius.all(
+                                    Radius.circular(12),
+                                  ),
+                                ),
+                                padding: const EdgeInsets.all(8),
+                                child: Column(
+                                  children: [
+                                    Icon(
+                                      Icons.video_call,
+                                      color: Theme.of(context).primaryColor,
+                                    ),
+                                    const SizedBox(
+                                      height: 6,
+                                    ),
+                                    Text(
+                                      "Escolha um vídeo",
+                                      style: TextStyle(
+                                        color: Theme.of(context).primaryColor,
+                                        fontSize: 12,
+                                      ),
+                                    ),
+                                  ],
+                                ),
+                              ),
+                            ],
+                          ),
                         ),
                         Center(
                           child: Text(L10N.of(context)!.helloWorld),
