+++ lib/src/utils/helpers/select_videos_sheet_helper.dart
@@ -1,0 +1,132 @@
+import 'dart:developer';
+
+import 'package:flutter/material.dart';
+
+import 'package:image_picker/image_picker.dart';
+import 'package:permission_handler/permission_handler.dart';
+
+import 'package:flutter_leap/src/utils/custom_colors.dart';
+import 'package:flutter_leap/src/utils/helpers/image_picker_helper.dart';
+
+void openSelectVideoSheet(BuildContext context) async {
+  try {
+    showModalBottomSheet(
+      backgroundColor: Colors.transparent,
+      context: context,
+      builder: (BuildContext context) {
+        return SafeArea(
+          child: SingleChildScrollView(
+            child: Container(
+              margin: const EdgeInsets.symmetric(horizontal: 10),
+              child: Wrap(
+                children: <Widget>[
+                  GestureDetector(
+                    onTap: () async {
+                      final status = await Permission.camera.request();
+                      if (status.isGranted) {
+                        getVideoHelper(source: ImageSource.camera);
+                        Navigator.pop(context);
+                      } else {
+                        Navigator.pop(context);
+                        ScaffoldMessenger.of(context).showSnackBar(
+                          const SnackBar(
+                            content: Text('Permissão de acesso não concedida'),
+                          ),
+                        );
+                      }
+                    },
+                    child: Container(
+                      width: double.infinity,
+                      height: 55,
+                      alignment: Alignment.center,
+                      decoration: BoxDecoration(
+                        color: CustomColors.white,
+                        borderRadius: const BorderRadius.only(
+                          topLeft: Radius.circular(15),
+                          topRight: Radius.circular(15),
+                        ),
+                      ),
+                      child: Text(
+                        "Gravar Vídeo",
+                        textAlign: TextAlign.center,
+                        style: TextStyle(
+                          color: CustomColors.black,
+                          fontSize: 16,
+                        ),
+                      ),
+                    ),
+                  ),
+                  GestureDetector(
+                    onTap: () async {
+                      final status = await Permission.storage.request();
+                      if (status.isGranted) {
+                        getVideoHelper(source: ImageSource.gallery);
+                        Navigator.pop(context);
+                      } else {
+                        Navigator.pop(context);
+                        ScaffoldMessenger.of(context).showSnackBar(
+                          const SnackBar(
+                            content: Text(
+                              'Permissão de acesso não concedida',
+                            ),
+                          ),
+                        );
+                      }
+                    },
+                    child: Container(
+                      width: double.infinity,
+                      height: 55,
+                      alignment: Alignment.center,
+                      decoration: BoxDecoration(
+                        color: CustomColors.white,
+                        borderRadius: const BorderRadius.only(
+                          bottomRight: Radius.circular(15),
+                          bottomLeft: Radius.circular(15),
+                        ),
+                      ),
+                      child: Text(
+                        "Buscar Vídeo",
+                        textAlign: TextAlign.center,
+                        style: TextStyle(
+                          color: CustomColors.black,
+                          fontSize: 16,
+                        ),
+                      ),
+                    ),
+                  ),
+                  GestureDetector(
+                    onTap: () => Navigator.pop(context),
+                    child: Container(
+                      margin: const EdgeInsets.only(top: 10, bottom: 30),
+                      width: double.infinity,
+                      height: 55,
+                      alignment: Alignment.center,
+                      decoration: BoxDecoration(
+                        color: CustomColors.white,
+                        borderRadius: const BorderRadius.all(
+                          Radius.circular(
+                            15,
+                          ),
+                        ),
+                      ),
+                      child: Text(
+                        "Cancelar",
+                        textAlign: TextAlign.center,
+                        style: TextStyle(
+                          color: CustomColors.black,
+                          fontSize: 16,
+                        ),
+                      ),
+                    ),
+                  ),
+                ],
+              ),
+            ),
+          ),
+        );
+      },
+    );
+  } catch (e) {
+    log(e.toString());
+  }
+}
